refactor: split the docstring line grammar out of the section walk
- Extracts `grammar.rs` from `section.rs`, holding the four line predicates the walk dispatches on
- Returns an `EntryHead` from `entry_head` and `sibling_entry_head` rather than an opaque pair
- Carries the head's colon offset on `SectionEntry`, so `docstring_sections` reads it rather than re-parsing
- Demotes `unbracketed_colon` to private, its last external consumer having stopped re-deriving the colon
